refactor: move the global client error code
