+++ packages/next/src/client/components/errors/console-error.ts
