+++ packages/next/src/client/components/errors/runtime-error-handler.ts
