+++ packages/next/src/client/components/globals/handle-global-errors.ts
@@ -1,3 +1,3 @@
-import { handleGlobalErrors } from '../react-dev-overlay/internal/helpers/use-error-handler'
+import { handleGlobalErrors } from '../errors/use-error-handler'
 
 handleGlobalErrors()
