+++ packages/next/src/client/components/react-dev-overlay/app/OldReactDevOverlay.tsx
@@ -9,7 +9,7 @@ import { ComponentStyles } from '../internal/styles/ComponentStyles'
 import { CssReset } from '../internal/styles/CssReset'
 import { RootLayoutMissingTagsError } from '../internal/container/RootLayoutMissingTagsError'
 import type { Dispatcher } from './hot-reloader-client'
-import { RuntimeErrorHandler } from '../internal/helpers/runtime-error-handler'
+import { RuntimeErrorHandler } from '../../errors/runtime-error-handler'
 
 interface ReactDevOverlayState {
   isReactError: boolean
