+++ packages/next/src/client/components/react-dev-overlay/pages/hot-reloader-client.ts
@@ -46,7 +46,7 @@ import type {
 } from '../../../../server/dev/hot-reloader-types'
 import { extractModulesFromTurbopackMessage } from '../../../../server/dev/extract-modules-from-turbopack-message'
 import { REACT_REFRESH_FULL_RELOAD_FROM_ERROR } from '../shared'
-import { RuntimeErrorHandler } from '../internal/helpers/runtime-error-handler'
+import { RuntimeErrorHandler } from '../../errors/runtime-error-handler'
 // This alternative WebpackDevServer combines the functionality of:
 // https://github.com/webpack/webpack-dev-server/blob/webpack-1/client/index.js
 // https://github.com/webpack/webpack/blob/webpack-1/hot/dev-server.js
